requirements.txt
asgiref==3.5.0
Django==3.2.11
djangorestframework==3.13.1
Pygments==2.11.2
pytz==2021.3
sqlparse==0.4.2